Refactor cartographer API into smaller modules
